+++ CHANGELOG.md
@@ -5,6 +5,13 @@ As mudanças feitas neste projeto serão documentadas nesse arquivo.
 O formato é baseado em [Mantenha uma Changelog](https://keepachangelog.com/en/1.0.0/),
 e este projeto segue [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 15-02-2022
+### Adicionado
+- Funcionalidade: Acessar e exibir modal do componente
+- Funcionalidade: Fechar plataforma
+### Mudado
+- Validação se já está na página da plataforma quando for testar outro cenário
+
 ## [1.0.0] - 14-02-2022
 ### Adicionado
 - Funcionalidade: Acessar plataforma IA